added separate db config handler component
