+++ storage/factory/export_test.go
@@ -2,16 +2,6 @@ package factory
 
 import "github.com/multiversx/mx-chain-go/config"
 
-// GetDBConfig -
-func (pf *PersisterFactory) GetDBConfig(path string) (*config.DBConfig, error) {
-	return pf.getDBConfig(path)
-}
-
-// CreatePersisterConfigFile -
-func CreatePersisterConfigFile(path string, dbConfig *config.DBConfig) error {
-	return createPersisterConfigFile(path, dbConfig)
-}
-
 // GetPersisterConfigFilePath -
 func GetPersisterConfigFilePath(path string) string {
 	return getPersisterConfigFilePath(path)
